% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig2tex.R
\name{fig2tex}
\alias{fig2tex}
\title{Save figure to TeX snippet}
\usage{
fig2tex(g, fileGraph, caption)
}
\arguments{
\item{g}{figure of class ggplot}

\item{caption}{caption for figure}

\item{filename}{file name to save figure}
}
\description{
Saves the figure and creates a small TeX snippet with caption and label, which can be
             compiled into a LaTeX document.
}
\examples{
q = tempdir()
x=seq(-10,10,0.1); y=x^2-2
g = ggplot2::ggplot(data.frame(x,y), ggplot2::aes(x,y)) +
  ggplot2::geom_point(col='red') + ggplot2::theme_bw()
fig2tex(g, file.path(q,"test.png"), "First Graph.")

}
