% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.fileHeader.R
\name{makeRAWtestFolder}
\alias{makeRAWtestFolder}
\title{Makes a RAW test folder}
\usage{
makeRAWtestFolder(noFiles = 1)
}
\arguments{
\item{noFiles}{number of random files to generate}
}
\description{
Makes a RAW test folder
}
