% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.goUpOneDir.R
\name{path.goUpOneDir}
\alias{path.goUpOneDir}
\title{Return path up one directory}
\usage{
path.goUpOneDir(p)
}
\arguments{
\item{p}{path}
}
\value{
path with last directory removed
}
\description{
Return path up one directory
}
\examples{
path.goUpOneDir('/Users/test/fish-100')
}
