% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.checkNoSubfolders.R
\name{raw.checkNoSubfolders}
\alias{raw.checkNoSubfolders}
\title{Checks that the RAW folder contains no sub-folders}
\usage{
raw.checkNoSubfolders(pfad)
}
\arguments{
\item{pfad}{path to the RAW folder}
}
\value{
boolean (true or false)
}
\description{
Checks that the RAW folder contains no sub-folders
}
\examples{
raw.checkNoSubfolders('.')

}
