% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.dataATE.R
\name{raw.dataATE}
\alias{raw.dataATE}
\title{Create dataATE Table}
\usage{
raw.dataATE(fIDfile = "data-raw/RAW-ID.csv", dataATE = NULL, verbose = FALSE)
}
\arguments{
\item{fIDfile}{RAW-ID file with path}

\item{dataATE}{table with the ATE data}

\item{verbose}{logical for extra information}
}
\description{
Create dataATE Table
}
