% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.dataFilesAFM.R
\name{raw.dataFilesAFM}
\alias{raw.dataFilesAFM}
\title{Create dataFilesAFM Table}
\usage{
raw.dataFilesAFM(fIDfile = "data-raw/RAW-ID.csv", dataFilesAFM = NULL)
}
\description{
Create dataFilesAFM Table
}
