% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.dataProject.R
\name{raw.dataProject}
\alias{raw.dataProject}
\title{makes a data R project}
\usage{
raw.dataProject(sourceDir = NULL)
}
\arguments{
\item{sourceDir}{if specified, R project will be generated in this directory, otherwise an interactive window allows user to select directory}
}
\description{
Use the interactive directory selector to create a data R project:
The selected folder should have a subdirectory called RAW and all
the RAW data files are stored in that directory
}
\examples{
\dontrun{
RAWdataR::raw.dataProject()
}
}
