% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.dataXRD.R
\name{raw.dataXRD}
\alias{raw.dataXRD}
\title{Create dataXRD Table}
\usage{
raw.dataXRD(fIDfile = "data-raw/RAW-ID.csv", dataXRD = NULL, verbose = FALSE)
}
\arguments{
\item{fIDfile}{RAW-ID file with path}

\item{verbose}{logical for extra information}

\item{dataFilesAFM}{table with the AFM data}
}
\description{
Create dataXRD Table
}
