% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.findDuplicates.R
\name{raw.findDuplicates}
\alias{raw.findDuplicates}
\title{Find duplicate RAW files}
\usage{
raw.findDuplicates(pfad, fromLast = TRUE, isValid = FALSE)
}
\arguments{
\item{pfad}{path to search for duplicate files}

\item{fromLast}{if \code{TRUE}, last element is kept, if \code{FALSE}, first element is kept, default is \code{TRUE}}

\item{isValid}{if \code{TRUE} only search valid named files}
}
\value{
list of files that are duplicates
}
\description{
Date: 2018-11-21
Searches a directory for all files (not necessarily coded properly), then returns a vector with files that have a duplicate.
}
