% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.findFiles.R
\name{raw.findFiles}
\alias{raw.findFiles}
\title{returns a list of files provided the search criteria}
\usage{
raw.findFiles(
  pfad,
  project = ".*",
  date = ".*",
  user = ".*",
  instrument = ".*",
  sample = ".*",
  md5 = ""
)
}
\arguments{
\item{project}{name of project}

\item{date}{date in format YYYYMMDD}

\item{user}{2 initials for user}

\item{instrument}{name of instrument, such as 'vsm'}

\item{sample}{sample name}

\item{file.list}{path of RAW files}
}
\value{
filename with certain checksum
}
\description{
returns a list of files provided the search criteria
}
\examples{
raw.findFiles(pfad, date='2020')  # all files from 2020

}
