% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.fixInvalidFile.R
\name{raw.fixInvalidFile}
\alias{raw.fixInvalidFile}
\title{Fix invalid filename}
\usage{
raw.fixInvalidFile(filename, instrument = "")
}
\arguments{
\item{filename}{name of file}

\item{instrument}{new name of instrument}
}
\value{
returns fixed filenames
}
\description{
Fix invalid filename
}
\examples{
f = raw.findFiles(raw.getSamplePath(), instrument='MM160622SI1')
print(f)
raw.fixInvalidFile(f, instrument='vsm')
}
