% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.generateDataMaker.R
\name{raw.generateDataMaker}
\alias{raw.generateDataMaker}
\title{creates a maker file inside the data-raw folder}
\usage{
raw.generateDataMaker(fileMaker = "make.dataRAW.R")
}
\description{
creates a maker file inside the data-raw folder
}
