% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.getDatabase.R
\name{raw.getDatabase}
\alias{raw.getDatabase}
\title{Manage SQL Database with Data Package}
\usage{
raw.getDatabase(pkgname, dbPath = NULL, verbose = FALSE)
}
\arguments{
\item{pkgname}{name of the R data package}

\item{dbPath}{path}

\item{verbose}{logical, additional information}
}
\value{
SQL database filename and path
}
\description{
Some data packages require large amounts of data (1 - 10GB), which cannot be
stored in the `extdata` folder directly, since the install would take too long.
If data exceeds about 10 MB, then RDA files become inefficient, and an SQL
database makes sense. For AFM data, you can use AFM.writeDB() for example.

The database needs to be stored in the `extdata` and also needs to be version
controlled. This function helps manage this process. The data package generates
the small RDA data files and puts the large data files into the SQL databse
that is stored in the main directory of the database. The database needs to store
at least one more dummy file in the `inst/extdata` folder, so that this folder
is generated and loaded.

When called with the `pkgname`, the function uses the version to generate the
database filename and return its path.
}
