% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.updateID.R
\name{raw.getFileByID}
\alias{raw.getFileByID}
\title{Return filename by ID}
\usage{
raw.getFileByID(ID, pRESULTS, idFile = "RAW-ID.csv")
}
\arguments{
\item{ID}{list of RAW file IDs}

\item{pRESULTS}{results folder}

\item{idFile}{name of the file the the RAW IDs}
}
\description{
Return filename by ID
}
\seealso{
\code{\link{raw.getIDbyFile}}, \code{\link{raw.updateID}}
}
