% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.getFilename.R
\name{raw.getFilename}
\alias{raw.getFilename}
\title{OBSOLETE: return filename from MD5}
\usage{
raw.getFilename(file.list, MD5)
}
\arguments{
\item{file.list}{path and filenames}

\item{MD5}{checksum, can be any length}
}
\description{
OBSOLETE: return filename from MD5
}
