% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.getFilename.R
\name{raw.getFilenameByMD5}
\alias{raw.getFilenameByMD5}
\title{returns the filename with certain MD5sum}
\usage{
raw.getFilenameByMD5(file.list, MD5)
}
\arguments{
\item{file.list}{path and filenames}

\item{MD5}{checksum, can be any length}
}
\value{
filename with certain MD5 checksum, returns \code{NA} if not found
}
\description{
returns the filename with certain MD5sum
}
\examples{
file.list = file.path(raw.getSamplePath(),dir(raw.getSamplePath()))
raw.getFilenameByMD5(file.list,'6b')
}
