% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.updateID.R
\name{raw.getIDbyFile}
\alias{raw.getIDbyFile}
\title{Return filename by ID}
\usage{
raw.getIDbyFile(file.list, pRESULTS, idFile = "RAW-ID.csv")
}
\arguments{
\item{file.list}{list of file names}

\item{pRESULTS}{results folder}

\item{idFile}{name of the file the the RAW IDs}
}
\description{
Return filename by ID
}
\seealso{
\code{\link{raw.getIDbyFile}}, \code{\link{raw.updateID}}
}
