% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.getPartialMD5.R
\name{raw.getMD5}
\alias{raw.getMD5}
\title{returns the MD5sum of a filename}
\usage{
raw.getMD5(filename, num = 6)
}
\arguments{
\item{filename}{filename}

\item{num}{string length}
}
\value{
CRC MD5 sum for this particular file
}
\description{
returns the MD5sum of a filename
}
\examples{
raw.getMD5(raw.getSampleFiles(),num=10)

}
