% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.getMD5.R
\name{raw.getMD5str}
\alias{raw.getMD5str}
\title{returns a string with all MD5 files}
\usage{
raw.getMD5str(filename)
}
\arguments{
\item{filename}{filename with path}
}
\value{
CRC MD5 sum for this particular file
}
\description{
returns a string with all MD5 files
}
\examples{
raw.getMD5str(raw.getSampleFiles())

}
