% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.inspectFolder.R
\name{raw.getNames}
\alias{raw.getNames}
\title{Separate projects, dates, and users}
\usage{
raw.getNames(file.list)
}
\arguments{
\item{file.list}{file names}
}
\value{
list
}
\description{
Separate projects, dates, and users
}
\examples{
raw.getNames(raw.getSampleFiles())
}
