% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.getNamesProjects.R
\name{raw.getNamesProjects}
\alias{raw.getNamesProjects}
\title{returns all project names}
\usage{
raw.getNamesProjects(pfad)
}
\arguments{
\item{pfad}{path to the RAW folder}
}
\value{
vector with project names
}
\description{
returns all project names
}
\examples{
raw.getNamesProjects('.')

}
