% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.getNoFiles.R
\name{raw.getNoFiles}
\alias{raw.getNoFiles}
\title{Number of Files}
\usage{
raw.getNoFiles(pfad)
}
\arguments{
\item{pfad}{path to the RAW folder}
}
\value{
number of files
}
\description{
returns number of files, but does not
include filenames that start with _, such
as _README.txt, which is not a data file
}
\examples{
raw.getNoFiles(raw.getSamplePath())

}
