% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.getNoImages.R
\name{raw.getNoImages}
\alias{raw.getNoImages}
\title{returns the number of images}
\usage{
raw.getNoImages(pfad)
}
\arguments{
\item{pfad}{path to the RAW folder}
}
\value{
number of images
}
\description{
returns the number of images
}
\examples{
raw.getNoImages(raw.getSamplePath())

}
