% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.getNoSamples.R
\name{raw.getNoSamples}
\alias{raw.getNoSamples}
\title{returns number of samples that were made}
\usage{
raw.getNoSamples(pfad)
}
\arguments{
\item{pfad}{path to the RAW folder}
}
\value{
number of samples that were made with the NTE
}
\description{
returns number of samples that were made
}
\examples{
raw.getNoSamples('.')

}
