% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.getMD5.R
\name{raw.getPartialMD5}
\alias{raw.getPartialMD5}
\title{returns the MD5sum of a filename}
\usage{
raw.getPartialMD5(filename, num = 6)
}
\arguments{
\item{filename}{filename}

\item{num}{string length}
}
\value{
CRC MD5 sum for this particular file
}
\description{
returns the MD5sum of a filename
}
