% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.getPartialMD5.R
\name{raw.getPartialMD5str}
\alias{raw.getPartialMD5str}
\title{returns a string with all MD5 files}
\usage{
raw.getPartialMD5str(filename)
}
\arguments{
\item{filename}{filename}

\item{num}{string length}
}
\value{
CRC MD5 sum for this particular file
}
\description{
returns a string with all MD5 files
}
\examples{
raw.getPartialMD5str('README.md')

}
