% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.getSamplePath.R
\name{raw.getSampleFiles}
\alias{raw.getSampleFiles}
\title{Sample Files from RAW folder}
\usage{
raw.getSampleFiles()
}
\value{
list of filenames including path
}
\description{
Returns a list of filenames to a sample RAW folder
}
\examples{
basename(raw.getSampleFiles())
}
