% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.getSamplePath.R
\name{raw.getSamplePath}
\alias{raw.getSamplePath}
\title{Sample path to RAW folder}
\usage{
raw.getSamplePath()
}
\value{
path with sample RAW files
}
\description{
Returns a path with several sample files to check validity of
filenames.
}
\examples{
\dontrun{
  raw.getSamplePath()
}
}
