% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.getUsername.R
\name{raw.getUsername}
\alias{raw.getUsername}
\title{returns the username from the path name}
\usage{
raw.getUsername(pfad)
}
\arguments{
\item{pfad}{path to the RAW folder}
}
\value{
username
}
\description{
returns the username from the path name
}
\examples{
raw.getUsername('Research-User')

}
