% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.getValidFiles.R
\name{raw.getValidFiles}
\alias{raw.getValidFiles}
\title{Returns all valid RAW files}
\usage{
raw.getValidFiles(pfad, recursive = FALSE)
}
\arguments{
\item{pfad}{path to the RAW folder}

\item{recursive}{if \code{TRUE}, will also search sub-directories}
}
\value{
vector with filenames
}
\description{
Searches a folder for valid files
}
\examples{
raw.getValidFiles(raw.getSamplePath())

}
