% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.installSQLremote.R
\name{raw.installSQLremote}
\alias{raw.installSQLremote}
\title{Install SQL AFM Database from Remote Site}
\usage{
raw.installSQLremote(pkgname, urlREPO)
}
\arguments{
\item{pkgname}{name of the package}

\item{urlREPO}{URL of directory with repository of SQL database}
}
\value{
path and file name of installed SQL database
}
\description{
Install SQL AFM Database from Remote Site
}
