% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.isValid.R
\name{raw.isValid}
\alias{raw.isValid}
\title{check whether a RAW data file has a valid filename}
\usage{
raw.isValid(filename)
}
\arguments{
\item{filename}{name of file}
}
\value{
returns \code{TRUE} if name has a valid format
}
\description{
check whether a RAW data file has a valid filename
}
\examples{
raw.isValid(raw.getSampleFiles())

}
