% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.RAWIDfile.R
\name{raw.RAWIDfile}
\alias{raw.RAWIDfile}
\title{Full Path of RAW-ID File}
\usage{
raw.RAWIDfile(pRESULTS = "data-raw", idFile = "RAW-ID.csv", ...)
}
\arguments{
\item{pRESULTS}{path for the RAW ID file}

\item{idFile}{name for the RAW-ID file, usually, this is RAW-ID.csv}

\item{...}{verbose logical}
}
\description{
The RAW-ID file is a CSV file with all the unique IDs and MD5 check sums
of the data files. This function guesses the location or returns it.
}
\examples{
##This may create a new directory
raw.RAWIDfile()

}
