% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.readRAWIDfile.R
\name{raw.readRAWIDfile}
\alias{raw.readRAWIDfile}
\title{Reads an RAW ID file}
\usage{
raw.readRAWIDfile(fIDfile = "data-raw/RAW-ID.csv")
}
\arguments{
\item{fIDfile}{file name, usually RAW_ID.csv}
}
\description{
Ignores header / comment lines with #
}
