% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.readRAWIDheader.R
\name{raw.readRAWIDheader}
\alias{raw.readRAWIDheader}
\title{Reads an RAW ID header}
\usage{
raw.readRAWIDheader(fIDfile)
}
\description{
Header information is preseded by # and separated name:value with colon
for example # Version: 1.0
}
\examples{
pRAW = makeRAWtestFolder()
raw.readRAWIDheader(file.path(pRAW, "RAW-ID.csv"))

}
