% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-duplicate-files.R
\name{raw.removeDuplicateFiles}
\alias{raw.removeDuplicateFiles}
\title{remove duplicate RAW files}
\usage{
raw.removeDuplicateFiles(pfad)
}
\arguments{
\item{pfad}{path to search for duplicate files}
}
\value{

}
\description{
remove duplicate RAW files
}
