% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.splitFilename.R
\name{raw.splitFilename}
\alias{raw.splitFilename}
\title{splits filename into fields}
\usage{
raw.splitFilename(filelist)
}
\arguments{
\item{filelist}{filename in RAW folder format}
}
\value{
data frame
}
\description{
splits filename into fields
}
\examples{
raw.splitFilename('20010101-project-user-inst-sample.DAT')

}
