% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.texArticle.R
\name{raw.texArticle}
\alias{raw.texArticle}
\title{LaTeX Article with all Fig2TeX figures}
\usage{
raw.texArticle(figPath, fileTeX = "_main.tex")
}
\arguments{
\item{figPath}{path for figures, should contain tex figure files generated with \code{fig2tex}}

\item{fileTex}{name of article that is generated in figPath}
}
\description{
LaTeX Article with all Fig2TeX figures
}
\seealso{
\code{\link{fig2tex()}}
}
