% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.updateID.R
\name{raw.updateID}
\alias{raw.updateID}
\title{Assigns a Unique ID to each RAW Data File}
\usage{
raw.updateID(
  pRAW = "",
  pRESULTS = "data-raw",
  idFile = "RAW-ID.csv",
  f_post = NA,
  forceRegenerate = FALSE,
  fixDuplicates = FALSE,
  removeIDs = c(),
  noData = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{pRAW}{path with raw data, if missing, then will prompt for path}

\item{pRESULTS}{path for results, default: uses pRAW}

\item{idFile}{name of file with IDs, default: RAW-ID.csv}

\item{f_post}{function to customize sample, type etc.}

\item{forceRegenerate}{logical, regenerate file, use with great care only}

\item{fixDuplicates}{logical, if \code{TRUE}, duplicates are removed, use with care only}

\item{removeIDs}{CAUTION: will delete IDs listed as vector}

\item{noData}{logical, if \code{TRUE}, returns RAW ID file name otherwise RAW data}

\item{verbose}{logical, if \code{TRUE} outputs information about the process}
}
\value{
returns \code{TRUE} if name has a valid format
}
\description{
Finds all RAW files and assigns a unique ID. Once the ID is assigned,
it is immutable; each file is identified by its CRC check sum,
code and file size, so not by the file name; if the file name or location
has changed, the ID will remain the same, but the file name is updated.
Some meta data with paths to RAW data is stored as header information, currently
located at the end of the file.
}
\section{Possible scenarios}{

\describe{
  \item{\code{(rule 1)}}{ RAW ID file does not exist and must be generated}
  \item{\code{(rule 2)}}{ new RAW file is added,}
  \item{\code{(rule 3)}}{ RAW file is missing,}
  \item{\code{(rule 4)}}{ RAW file has duplicate in different folder,}
  \item{\code{(rule 5)}}{ RAW file has duplicate with new name,}
  \item{\code{(rule 6)}}{ RAW file moved to a different folder, and folder needs update}
}
}

\examples{
\dontrun{
  raw.updateID()
}
}
\seealso{
\code{\link{raw.getFileByID}}, \code{\link{raw.getIDbyFile}}
}
