% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.writeRAWIDfile.R
\name{raw.writeRAWIDfile}
\alias{raw.writeRAWIDfile}
\title{Writes an RAW ID file with Header}
\usage{
raw.writeRAWIDfile(rID, header_list, fIDfile)
}
\arguments{
\item{rID}{RAW ID data frame}

\item{header_list}{list to be saved}

\item{fIDfile}{filen ame RAW_ID.csv to be saved}
}
\description{
Writes an RAW ID file with Header
}
