% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.writeRAWIDfile.R
\name{raw.writeRAWIDheader}
\alias{raw.writeRAWIDheader}
\title{Writes a RAW ID Header}
\usage{
raw.writeRAWIDheader(header_list, fIDfile)
}
\arguments{
\item{header_list}{list to be saved}

\item{fIDfile}{filen ame RAW_ID.csv to be saved}
}
\description{
Writes a RAW ID Header
}
