% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.checkNoSubfolders.R
\name{raw.checkNoSubfolders}
\alias{raw.checkNoSubfolders}
\title{Subfolders in RAW path}
\usage{
raw.checkNoSubfolders(pfad)
}
\arguments{
\item{pfad}{path to the RAW folder}
}
\value{
\code{TRUE} if there are no subfolders
}
\description{
The RAW path should not contain subfolders; the reason for
this is that subfolders tend to have additional information in
the folder name, which then makes the file less unique.
}
\examples{
raw.checkNoSubfolders(raw.getSamplePath())
}
