% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.dataMaker.R
\name{raw.dataMaker}
\alias{raw.dataMaker}
\title{creates a maker file inside the data-raw folder}
\usage{
raw.dataMaker(fileMaker = NULL)
}
\arguments{
\item{fileMaker}{type of file to generate, if \code{NULL} options are provided via terminal}
}
\description{
The
}
\examples{
\donotrun{
raw.dataMaker()
}
}
