% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.dataMaker.R
\name{raw.dataMaker}
\alias{raw.dataMaker}
\title{Creates a maker file inside the data-raw folder}
\usage{
raw.dataMaker(fileMaker = NULL)
}
\arguments{
\item{fileMaker}{type of file to generate, if \code{NULL} options are provided via terminal}
}
\description{
Creates template files that help create data sets; several
data set makers are supported; use no arguments to see the
interactive menu.
}
\examples{
\dontrun{
raw.dataMaker()
}
}
