% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.findFiles.R
\name{raw.findFiles}
\alias{raw.findFiles}
\title{returns a list of files provided the search criteria}
\usage{
raw.findFiles(
  pfad,
  project = ".*",
  date = ".*",
  user = ".*",
  instrument = ".*",
  sample = ".*",
  md5 = ""
)
}
\arguments{
\item{project}{name of project}

\item{date}{date in format YYYYMMDD}

\item{user}{2 initials for user}

\item{instrument}{name of instrument, such as 'vsm'}

\item{sample}{sample name}

\item{md5}{single string with comma separated abbreviated MD5 sums}

\item{file.list}{path of RAW files}
}
\value{
filename with certain checksum
}
\description{
returns a list of files provided the search criteria
}
\examples{
file.list = raw.findFiles(pfad, date='201901')  # all files from Jan 2019
md5String = raw.getPartialMD5str(file.list)
file.list = raw.findFiles(pfad, date='201901',md5 = md5String)

}
