% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.fixInvalidFile.R
\name{raw.fixInvalidFile}
\alias{raw.fixInvalidFile}
\title{Fix invalid data filename}
\usage{
raw.fixInvalidFile(
  filename,
  addInstrument = "",
  addUser = "",
  addProject = "",
  addSample = "",
  renameFile = FALSE
)
}
\arguments{
\item{filename}{name of file}

\item{addInstrument}{instrument to be added}

\item{addUser}{user to be added}

\item{addProject}{project to be added}

\item{addSample}{sample to be added}

\item{renameFile}{if \code{TRUE} file will be renamed}
}
\value{
returns fixed filenames
}
\description{
Fix invalid data filename
}
\examples{
f = raw.findFiles(raw.getSamplePath(), addInstrument='MM160622SI1')
print(f)
f.new = raw.fixInvalidFile(f, instrument='vsm')
print(paste("Fixing:",f,"to new file:",f.new))
}
