% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.getFilename.R
\name{raw.getFilename}
\alias{raw.getFilename}
\title{returns the filename with certain MD5sum}
\usage{
raw.getFilename(file.list, MD5)
}
\arguments{
\item{file.list}{path and filenames}
}
\value{
filename with certain checksum
}
\description{
returns the filename with certain MD5sum
}
\examples{
raw.getFilename(dir(),'6')

}
