% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.updateID.R
\name{raw.getIDbyFile}
\alias{raw.getIDbyFile}
\title{Return filename by ID}
\usage{
raw.getIDbyFile(
  file.list,
  pRESULTS = "data-raw",
  idFile = "RAW-ID.csv",
  exactNameMatch = TRUE
)
}
\arguments{
\item{file.list}{list of file names}

\item{pRESULTS}{results folder, if missing and path.RESULTS is defined will use that folder}

\item{idFile}{name of the file the the RAW IDs}

\item{exactNameMatch}{logical, if \code{FALSE} will match partial file names}
}
\description{
Return filename by ID
}
\seealso{
\code{\link{raw.getIDbyFile}}, \code{\link{raw.updateID}}
}
