% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.getValidFiles.R
\name{raw.getInvalidFiles}
\alias{raw.getInvalidFiles}
\title{Returns all invalid RAW files}
\usage{
raw.getInvalidFiles(pfad, year = ".*", recursive = FALSE)
}
\arguments{
\item{pfad}{path to the RAW folder}

\item{year}{limit to a particular year (currently ignored)}

\item{recursive}{if \code{TRUE}, will also search sub-directories}
}
\value{
vector with filenames
}
\description{
Searches a folder (and year) for files that could be invalid
}
\examples{
raw.getInvalidFiles(raw.getSamplePath())

}
