% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.getValidFiles.R
\name{raw.getInvalidFiles}
\alias{raw.getInvalidFiles}
\title{returns all invalid RAW files in a folder}
\usage{
raw.getInvalidFiles(pfad, year = ".*")
}
\arguments{
\item{pfad}{path to the RAW folder}

\item{year}{4-digit year to limit search}
}
\value{
vector with filenames
}
\description{
returns all invalid RAW files in a folder
}
\examples{
raw.getInvalidFiles('.')

}
