% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.getNoSamples.R
\name{raw.getNoSamples}
\alias{raw.getNoSamples}
\title{returns approximate number of samples that were made}
\usage{
raw.getNoSamples(pfad, underscoreComments = TRUE)
}
\arguments{
\item{pfad}{path to the RAW folder}

\item{underscoreComments}{include files with comments that have _}
}
\value{
number of samples that were made with the NTE
}
\description{
returns approximate number of samples that were made
}
\examples{
raw.getNoSamples(raw.getSamplePath())

}
