% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.getTable.R
\name{raw.getTable}
\alias{raw.getTable}
\title{returns a dataframe with field information for each file}
\usage{
raw.getTable(pfad, project = "")
}
\arguments{
\item{pfad}{path to the RAW folder}

\item{project}{project name limits files}
}
\value{
dataframe
}
\description{
returns a dataframe with field information for each file
}
\examples{
raw.getTable('.')

}
