% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.getValidFiles.R
\name{raw.getValidFiles}
\alias{raw.getValidFiles}
\title{returns all valid RAW files in a folder}
\usage{
raw.getValidFiles(pfad)
}
\arguments{
\item{pfad}{path to the RAW folder}
}
\value{
vector with filenames
}
\description{
returns all valid RAW files in a folder
}
\examples{
raw.getValidFiles('.')

}
