% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.inspectFolder.R
\name{raw.inspectFolder}
\alias{raw.inspectFolder}
\title{Inspect RAW data folder}
\usage{
raw.inspectFolder(pfad)
}
\arguments{
\item{pfad}{path to folder with RAW data}
}
\value{
list
}
\description{
Inspect RAW data folder
}
\examples{
raw.inspectFolder(raw.getSamplePath())
}
