% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.installSQLremote.R
\name{raw.installSQLremote}
\alias{raw.installSQLremote}
\title{Install SQL AFM Database from Remote Site}
\usage{
raw.installSQLremote(pkgname, urlRepo, upgradeFrom = "")
}
\arguments{
\item{pkgname}{file name of the SQL db package}

\item{upgradeFrom}{old version, string such as "0.2.1", black if not an upgrade}

\item{urlREPO}{URL of directory with repository of SQL database}
}
\value{
SQL database filename and path
}
\description{
Installs a large SQL AFM database from a remote site; in order
for data R packages to load more quickly and not take up large
space on external repositories, this function implements a solution
that stores the database remotely. When needed, it will be installed
locally in the `extdata/` folder. The function returns the name of
the database.

If new data is added to the SQL databse, a new version is generally
generated; this can be done by upgrading from an existing old version.
}
