% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.installSQLremote.R
\name{raw.installSQLremote}
\alias{raw.installSQLremote}
\title{Install SQL AFM Database from Remote Site}
\usage{
raw.installSQLremote(dbName, targetDir, urlRepo)
}
\arguments{
\item{dbName}{file name of the SQL db package}

\item{targetDir}{target directory to install, usually `inst/extdata`}

\item{urlREPO}{URL of directory with repository of SQL database}
}
\value{
logical, \code{TRUE}, if SQL db exists or was successfully installed

logical, if \code{TRUE} DB is available
}
\description{
Installs a large SQL AFM database from a remote site; in order
for data R packages to load more quickly and not take up large
space on GitHub; it can separately store the AFM images in a
SQL database, which can usually be the size of about 1 GB. This
database is stored in the `inst/extdata` data folder. First,
create an empty SQL database file in the inst/extdata folder
}
