% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.texArticle.R
\name{raw.texArticle}
\alias{raw.texArticle}
\title{LaTeX Article with all Fig2TeX figures}
\usage{
raw.texArticle(figPath, fileTeX = "_main.tex")
}
\arguments{
\item{figPath}{path for figures, should contain tex figure files generated with \code{fig2tex}}

\item{fileTeX}{name of article that is generated in figPath}
}
\description{
After saving TeX snippets into a folder (\code{figPath}) with \code{\link{fig2tex}},
then, use this function to generate
a main TeX file called \code{fileTeX} that includes all the snippets and after
compiling shows a document with all the graphs / images. The main TeX file is
saved in the same folder as the \code{figPath}. It will overwrite a file.
}
\seealso{
\code{\link{fig2tex}}
}
