% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.updateID.R
\name{raw.updateID}
\alias{raw.updateID}
\title{Assigns ID to each RAW data file}
\usage{
raw.updateID(
  pRAW,
  pRESULTS,
  idFile = "RAW-ID.csv",
  forceRegenerate = FALSE,
  fixDuplicates = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{pRAW}{path with raw data, default: uses `data-raw` folder, or creates it}

\item{pRESULTS}{path for results, default: uses pRAW}

\item{idFile}{name of file with IDs, default: RAW-ID.csv}

\item{forceRegenerate}{logical, regenerate file, use with great care only}

\item{fixDuplicates}{logical, if \code{TRUE}, duplicates are removed, use with care only}

\item{verbose}{logical, if \code{TRUE} outputs information about the process}
}
\value{
returns \code{TRUE} if name has a valid format
}
\description{
Code finds all RAW files and assigns an ID, once the ID is assigned,
it should stay with that file; the file is identified by the CRC checksum
code and filesize, so not necessarily by the filename; if the filename
has changed, the ID will remain the same, but the file is changing
}
\section{Possible scenarios}{

\describe{
  \item{\code{(rule 1)}}{ RAW ID file does not exist and must be generated}
  \item{\code{(rule 2)}}{ new RAW file is added,}
  \item{\code{(rule 3)}}{ RAW file is missing,}
  \item{\code{(rule 4)}}{ RAW file has duplicate in different folder,}
  \item{\code{(rule 5)}}{ RAW file has duplicate with new name,}
  \item{\code{(rule 6)}}{ RAW file moved to a different folder, and folder needs update}
}
}

\examples{
\dontrun{
  raw.updateID()
}
}
\seealso{
\code{\link{raw.getFileByID}}, \code{\link{raw.getIDbyFile}}
}
