% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.updateID.R
\name{raw.updateID}
\alias{raw.updateID}
\title{Assigns ID to each RAW data file}
\usage{
raw.updateID(
  pRAW,
  pRESULTS,
  idFile = "RAW-ID.csv",
  forceRegenerate = FALSE,
  fixDuplicates = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{pRAW}{path with raw data}

\item{pRESULTS}{path for results}

\item{idFile}{name of file with IDs}

\item{forceRegenerate}{logical, regenerate file, use with great care only}

\item{fixDuplicates}{logical, if \code{TRUE}, duplicates are removed, use with care only}

\item{verbose}{logical, if \code{TRUE} outputs information about the process}
}
\value{
returns \code{TRUE} if name has a valid format
}
\description{
Code finds all RAW files and assigns an ID, once the ID is assigned,
it should stay with that file; the file is identified by the CRC checksum
code and filesize, so not necessarily by the filename; if the filename
has changed, the ID will remain the same, but the file is changing
}
\section{Possible scenarios}{

(1) RAW ID file does not exist and must be generated
(2) new RAW file is added,
(3) RAW file is missing,
(4) RAW file has duplicate in different folder,
(5) RAW file has duplicate with new name,
(6) RAW file moved to a different folder, and folder needs update
}

\examples{
if (FALSE) raw.updateID(path.RAW, path.RESULTS)

}
